Split main feed/item layout into two separate components -- one for default layout and the other for ticker
